requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
blinker==1.7.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.104.1
flask==2.3.3
flask-restx==1.2.0
flask-sqlalchemy==3.1.1
greenlet==3.0.1
h11==0.14.0
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
pkgutil-resolve-name==1.3.10
protobuf==4.21.12
pydantic==2.5.2
pydantic-core==2.14.5
PyJWT==2.8.0
pyrsistent==0.20.0
pytz==2023.3.post1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing-extensions==4.8.0
uvicorn==0.24.0.post1
werkzeug==2.3.8
zipp==3.17.0
